Refactor chunkserver to generate layers of snapshots
A chunk might have multiple snapshots when we remove the assumption
that snapshots are always exported and deleted afterwards.

Changes are:
1. Introduce CSSnapshots to manage snapshots;
2. `correctedSn' no longer used for CoW logic.
3. Two new fields are introduced to `ChunkRequest':
  - snapSn, the snapshot sequence number;
  - snaps, all snapshot sequence numbers.

TODO:
1. managing of `snaps' field in client side.
2. CRUD of `internal' snapshots.
3. storage consumption and statistics - take care of CoW files.
4. Recover & clone of `internal' snapshots.
5. caching of meta objects, and take care of memory footprint.
6. optimize deletion - no need to merge if intended to delete both file and snapshots.
7. Remove `correctedSn' from `ChunkFileMetaPage' and 'CSChunkInfo'.

UT:
1. `test/chunkserver/datastore/curve_datastore_unittest' still fails.
2. new unit tests is still under development.


fix all ut failure in chunkserver/datastore

1. fix bitmap wrong size bug in CSSnapshots::Read method
2. Deletion of snapshot is only allowed when chunk.sn larger than latest snapshot sn
3. fix ut failure


Add ut case for local multi-snapshot feature

1. add manual ut case for multi-snapshot datastore and all passed
2. fix some bug

Support multi-snapshot in client side


Support read snapshot in curve client for the purpose of snapshot consistency check


Add CreateSyncSnapshot method in snapshotclone and multi-snapshot consistency check tool


Code improvement about client read-snapshot implementation 1. handle cases where snapshot sequence num in invalid 2. do not call IncreaseEpoch for snapshot file 3. other stuff


Fix nebd-part1 block during exit


eliminate the requirement of S3 store

(cherry picked from commit 28751eab2cd56761fd5a01d6710a96e56cfae192)

Support multi-level snapshot deletion orderly


Fix ReadChunkSnap bug

Implement OnChunkNotExist for ReadChunkSnapClosure to handle cases where chunk does not exist in chunkserver

+++ src/client/libcurve_file.cpp
@@ -184,6 +184,7 @@ int FileClient::Open(const std::string& filename,
                      const UserInfo_t& userinfo,
                      const OpenFlags& openflags) {
     LOG(INFO) << "Opening filename: " << filename << ", flags: " << openflags;
+
     ClientConfig clientConfig;
     if (openflags.confPath.empty()) {
         clientConfig = clientconfig_;
@@ -202,17 +203,28 @@ int FileClient::Open(const std::string& filename,
         return -LIBCURVE_ERROR::FAILED;
     }
 
-    FileInstance *fileserv = FileInstance::NewInitedFileInstance(
-        clientConfig.GetFileServiceOption(), mdsClient, filename, userinfo,
-        openflags, false);
+    uint64_t sn = 0;
+    std::string realfilename = filename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if (isSnapshot) {
+        LOG(INFO) << "Opening Snapshot sn: " << sn << ", realfilename: " << realfilename;
+    }
+
+    FileInstance* fileserv = FileInstance::NewInitedFileInstance(
+        clientConfig.GetFileServiceOption(), mdsClient, realfilename, userinfo,
+        openflags, isSnapshot);
     if (fileserv == nullptr) {
         LOG(ERROR) << "NewInitedFileInstance fail";
         return -1;
     }
 
+    if (isSnapshot) {
+        fileserv->SetReadSnapshotSn(sn);
+    }
+
     int ret = fileserv->Open();
     if (ret != LIBCURVE_ERROR::OK) {
-        LOG(ERROR) << "Open file failed, filename: " << filename
+        LOG(ERROR) << "Open file failed, filename: " << realfilename
                    << ", retCode: " << ret;
         fileserv->UnInitialize();
         delete fileserv;
@@ -224,10 +236,11 @@ int FileClient::Open(const std::string& filename,
     {
         WriteLockGuard lk(rwlock_);
         fileserviceMap_[fd] = fileserv;
-        fileserviceFileNameMap_[filename] = fileserv;
+        if (!isSnapshot)  // As there is no need for snapshot to IncreaseEpoch
+            fileserviceFileNameMap_[realfilename] = fileserv;
     }
 
-    LOG(INFO) << "Open success, filname = " << filename << ", fd = " << fd;
+    LOG(INFO) << "Open success, filname = " << realfilename << ", fd = " << fd;
     openedFileNum_ << 1;
 
     return fd;
@@ -235,6 +248,13 @@ int FileClient::Open(const std::string& filename,
 
 int FileClient::Open4ReadOnly(const std::string& filename,
                               const UserInfo_t& userinfo, bool disableStripe) {
+    uint64_t sn = 0;
+    std::string realfilename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if (isSnapshot) {
+        LOG(INFO) << "Open4ReadOnly Snapshot not allowed, sn: " << sn << ", realfilename: " << realfilename;
+        return -LIBCURVE_ERROR::NOT_SUPPORT;
+    }
     FileInstance* instance = FileInstance::Open4Readonly(
         clientconfig_.GetFileServiceOption(), mdsClient_, filename, userinfo);
 
@@ -263,6 +283,14 @@ int FileClient::Open4ReadOnly(const std::string& filename,
 int FileClient::IncreaseEpoch(const std::string& filename,
                               const UserInfo_t& userinfo) {
     LOG(INFO) << "IncreaseEpoch, filename: " << filename;
+    uint64_t sn = 0;
+    std::string realfilename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if (isSnapshot) {
+        LOG(INFO) << "IncreaseEpoch Snapshot not allowed, sn: " << sn << ", realfilename: " << realfilename;
+        return -LIBCURVE_ERROR::NOT_SUPPORT;
+    } 
+
     FInfo_t fi;
     FileEpoch_t fEpoch;
     std::list<CopysetPeerInfo> csLocs;
@@ -416,6 +444,13 @@ int FileClient::Rename(const UserInfo_t& userinfo,
 int FileClient::Extend(const std::string& filename,
     const UserInfo_t& userinfo, uint64_t newsize) {
     LIBCURVE_ERROR ret;
+    uint64_t sn = 0;
+    std::string realfilename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if (isSnapshot) {
+        LOG(INFO) << "Extend Snapshot not allowed, sn: " << sn << ", realfilename: " << realfilename;
+        return -LIBCURVE_ERROR::NOT_SUPPORT;
+    }  
     if (mdsClient_ != nullptr) {
         ret = mdsClient_->Extend(filename, userinfo, newsize);
         LOG_IF(ERROR, ret != LIBCURVE_ERROR::OK)
@@ -432,6 +467,13 @@ int FileClient::Extend(const std::string& filename,
 int FileClient::Unlink(const std::string& filename,
     const UserInfo_t& userinfo, bool deleteforce) {
     LIBCURVE_ERROR ret;
+    uint64_t sn = 0;
+    std::string realfilename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if (isSnapshot) {
+        LOG(INFO) << "Unlink Snapshot not allowed, sn: " << sn << ", realfilename: " << realfilename;
+        return -LIBCURVE_ERROR::NOT_SUPPORT;
+    }  
     if (mdsClient_ != nullptr) {
         ret = mdsClient_->DeleteFile(filename, userinfo, deleteforce);
         LOG_IF(ERROR, ret != LIBCURVE_ERROR::OK)
@@ -482,10 +524,16 @@ int FileClient::StatFile(const std::string &filename,
     FInfo_t fi;
     FileEpoch_t fEpoch;
     int ret;
+    uint64_t sn = 0;
+    std::string realfilename = filename;
+    bool isSnapshot = curve::client::ServiceHelper::GetSnapSeqFromFilename(filename, sn, &realfilename);
+    if (isSnapshot) {
+        LOG(INFO) << "StatFile Snapshot sn: " << sn << ", realfilename: " << realfilename;
+    }    
     if (mdsClient_ != nullptr) {
-        ret = mdsClient_->GetFileInfo(filename, userinfo, &fi, &fEpoch);
+        ret = mdsClient_->GetFileInfo(realfilename, userinfo, &fi, &fEpoch);
         LOG_IF(ERROR, ret != LIBCURVE_ERROR::OK)
-            << "StatFile failed, filename: " << filename << ", ret" << ret;
+            << "StatFile failed, filename: " << realfilename << ", ret" << ret;
     } else {
         LOG(ERROR) << "global mds client not inited!";
         return -LIBCURVE_ERROR::FAILED;
